Mudei os tipos das variaveis

+++ app/src/main/java/com/example/calculadora_reviso/Calculadora.kt
@@ -1,24 +1,23 @@
 package com.example.calculadora_reviso
 
+//essa classe executa o calculo
 class Calculadora {
 
-    fun somar(N1 : Int, N2 : Int): Int{
+    fun somar(N1 : Float, N2 : Float): Float{
         val result =  N1 + N2
         return(result)
     }
 
-    fun menos(N1 : Int, N2 : Int): Int{
+    fun menos(N1 : Float, N2 : Float): Float{
         val result = N1 - N2
         return(result)
     }
 
-    fun multiplicar(N1 : Int, N2 : Int): Int
-    {
+    fun multiplicar(N1 : Float, N2 : Float): Float{
         val result = N1 * N2
         return(result)
     }
-    fun divisao(N1 : Int, N2 : Int): Int
-    {
+    fun divisao(N1 : Float, N2 : Float): Float{
         val result = N1 / N2
         return(result)
     }
